pass method names for NFT API
