+++ src/api/nft-namespace.ts
@@ -1,4 +1,4 @@
-import { BigNumber, BigNumberish } from '@ethersproject/bignumber';
+import type { BigNumberish } from '@ethersproject/bignumber';
 import {
   GetBaseNftsForContractOptions,
   GetBaseNftsForOwnerOptions,
@@ -14,33 +14,26 @@ import {
   OwnedBaseNftsResponse,
   OwnedNft,
   OwnedNftsResponse,
-  RefreshContractResult,
-  RefreshState
+  RefreshContractResult
 } from '../types/types';
 import { BaseNft, Nft, NftContract } from './nft';
 import { AlchemyConfig } from './alchemy-config';
-import { paginateEndpoint, requestHttpWithBackoff } from '../internal/dispatch';
 import {
-  RawBaseNft,
-  RawGetBaseNftsForContractResponse,
-  RawGetBaseNftsResponse,
-  RawGetNftsForContractResponse,
-  RawGetNftsResponse,
-  RawGetOwnersForContractResponse,
-  RawNft,
-  RawNftContract,
-  RawContractBaseNft,
-  RawContractNft,
-  RawOwnedBaseNft,
-  RawOwnedNft,
-  RawReingestContractResponse
-} from '../internal/raw-interfaces';
-import { AlchemyApiType } from '../util/const';
-import {
-  getBaseNftFromRaw,
-  getNftContractFromRaw,
-  getNftFromRaw
-} from '../util/util';
+  checkNftOwnership,
+  getContractMetadata,
+  getFloorPrice,
+  getNftMetadata,
+  getNftsForContract,
+  getNftsForContractIterator,
+  getNftsForOwner,
+  getNftsForOwnerIterator,
+  getOwnersForContract,
+  getOwnersForNft,
+  getSpamContracts,
+  isSpamContract,
+  refreshContract,
+  refreshNftMetadata
+} from '../internal/nft-api';
 
 export class NftNamespace {
   constructor(private readonly config: AlchemyConfig) {}
@@ -53,22 +46,12 @@ export class NftNamespace {
    * @param tokenType - Optionally specify the type of token to speed up the query.
    * @public
    */
-  async getNftMetadata(
+  getNftMetadata(
     contractAddress: string,
     tokenId: BigNumberish,
     tokenType?: NftTokenType
   ): Promise<Nft> {
-    const response = await requestHttpWithBackoff<GetNftMetadataParams, RawNft>(
-      this.config,
-      AlchemyApiType.NFT,
-      'getNFTMetadata',
-      {
-        contractAddress,
-        tokenId: BigNumber.from(tokenId!).toString(),
-        tokenType: tokenType !== NftTokenType.UNKNOWN ? tokenType : undefined
-      }
-    );
-    return getNftFromRaw(response, contractAddress);
+    return getNftMetadata(this.config, contractAddress, tokenId, tokenType);
   }
 
   /**
@@ -77,15 +60,8 @@ export class NftNamespace {
    * @param contractAddress - The contract address of the NFT.
    * @public
    */
-  async getContractMetadata(contractAddress: string): Promise<NftContract> {
-    const response = await requestHttpWithBackoff<
-      GetContractMetadataParams,
-      RawNftContract
-    >(this.config, AlchemyApiType.NFT, 'getContractMetadata', {
-      contractAddress
-    });
-
-    return getNftContractFromRaw(response);
+  getContractMetadata(contractAddress: string): Promise<NftContract> {
+    return getContractMetadata(this.config, contractAddress);
   }
 
   /**
@@ -116,34 +92,11 @@ export class NftNamespace {
     owner: string,
     options?: GetBaseNftsForOwnerOptions
   ): AsyncIterable<OwnedBaseNft>;
-  async *getNftsForOwnerIterator(
+  getNftsForOwnerIterator(
     owner: string,
     options?: GetNftsForOwnerOptions | GetBaseNftsForOwnerOptions
   ): AsyncIterable<OwnedBaseNft | OwnedNft> {
-    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
-    for await (const response of paginateEndpoint(
-      this.config,
-      AlchemyApiType.NFT,
-      'getNFTs',
-      'pageKey',
-      'pageKey',
-      {
-        contractAddresses: options?.contractAddresses,
-        pageKey: options?.pageKey,
-        filters: options?.excludeFilters,
-        owner,
-        withMetadata
-      }
-    )) {
-      for (const ownedNft of response.ownedNfts as
-        | RawOwnedNft[]
-        | RawOwnedBaseNft[]) {
-        yield {
-          ...nftFromGetNftResponse(ownedNft),
-          balance: parseInt(ownedNft.balance)
-        };
-      }
-    }
+    return getNftsForOwnerIterator(this.config, owner, options);
   }
 
   /**
@@ -174,29 +127,11 @@ export class NftNamespace {
     owner: string,
     options?: GetBaseNftsForOwnerOptions
   ): Promise<OwnedBaseNftsResponse>;
-  async getNftsForOwner(
+  getNftsForOwner(
     owner: string,
     options?: GetNftsForOwnerOptions | GetBaseNftsForOwnerOptions
   ): Promise<OwnedNftsResponse | OwnedBaseNftsResponse> {
-    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
-    const response = await requestHttpWithBackoff<
-      GetNftsAlchemyParams,
-      RawGetBaseNftsResponse | RawGetNftsResponse
-    >(this.config, AlchemyApiType.NFT, 'getNFTs', {
-      contractAddresses: options?.contractAddresses,
-      pageKey: options?.pageKey,
-      filters: options?.excludeFilters,
-      owner,
-      withMetadata
-    });
-    return {
-      ownedNfts: response.ownedNfts.map(res => ({
-        ...nftFromGetNftResponse(res),
-        balance: parseInt(res.balance)
-      })),
-      pageKey: response.pageKey,
-      totalCount: response.totalCount
-    };
+    return getNftsForOwner(this.config, owner, options);
   }
 
   /**
@@ -228,26 +163,11 @@ export class NftNamespace {
     contractAddress: string,
     options?: GetBaseNftsForContractOptions
   ): Promise<NftContractBaseNftsResponse>;
-  async getNftsForContract(
+  getNftsForContract(
     contractAddress: string,
     options?: GetBaseNftsForContractOptions | GetNftsForContractOptions
   ): Promise<NftContractNftsResponse | NftContractBaseNftsResponse> {
-    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
-    const response = await requestHttpWithBackoff<
-      GetNftsForNftContractAlchemyParams,
-      RawGetBaseNftsForContractResponse | RawGetNftsForContractResponse
-    >(this.config, AlchemyApiType.NFT, 'getNFTsForCollection', {
-      contractAddress,
-      startToken: options?.pageKey,
-      withMetadata
-    });
-
-    return {
-      nfts: response.nfts.map(res =>
-        nftFromGetNftNftContractResponse(res, contractAddress)
-      ),
-      pageKey: response.nextToken
-    };
+    return getNftsForContract(this.config, contractAddress, options);
   }
 
   /**
@@ -281,29 +201,11 @@ export class NftNamespace {
     contractAddress: string,
     options?: GetBaseNftsForContractOptions
   ): AsyncIterable<BaseNft>;
-  async *getNftsForContractIterator(
+  getNftsForContractIterator(
     contractAddress: string,
     options?: GetBaseNftsForContractOptions | GetNftsForContractOptions
   ): AsyncIterable<BaseNft | Nft> {
-    const withMetadata = omitMetadataToWithMetadata(options?.omitMetadata);
-    for await (const response of paginateEndpoint(
-      this.config,
-      AlchemyApiType.NFT,
-      'getNFTsForCollection',
-      'startToken',
-      'nextToken',
-      {
-        contractAddress,
-        startToken: options?.pageKey,
-        withMetadata
-      }
-    )) {
-      for (const nft of response.nfts as
-        | RawContractBaseNft[]
-        | RawContractNft[]) {
-        yield nftFromGetNftNftContractResponse(nft, contractAddress);
-      }
-    }
+    return getNftsForContractIterator(this.config, contractAddress, options);
   }
 
   /**
@@ -312,19 +214,10 @@ export class NftNamespace {
    * @param contractAddress - The NFT contract to get the owners for.
    * @beta
    */
-  async getOwnersForContract(
+  getOwnersForContract(
     contractAddress: string
   ): Promise<GetOwnersForContractResponse> {
-    const response = await requestHttpWithBackoff<
-      GetOwnersForNftContractAlchemyParams,
-      RawGetOwnersForContractResponse
-    >(this.config, AlchemyApiType.NFT, 'getOwnersForCollection', {
-      contractAddress
-    });
-
-    return {
-      owners: response.ownerAddresses
-    };
+    return getOwnersForContract(this.config, contractAddress);
   }
 
   /**
@@ -334,19 +227,11 @@ export class NftNamespace {
    * @param tokenId - Token id of the NFT.
    * @beta
    */
-  async getOwnersForNft(
+  getOwnersForNft(
     contractAddress: string,
     tokenId: BigNumberish
   ): Promise<GetOwnersForNftResponse> {
-    return requestHttpWithBackoff(
-      this.config,
-      AlchemyApiType.NFT,
-      'getOwnersForToken',
-      {
-        contractAddress,
-        tokenId: BigNumber.from(tokenId!).toString()
-      }
-    );
+    return getOwnersForNft(this.config, contractAddress, tokenId);
   }
 
   /**
@@ -356,18 +241,11 @@ export class NftNamespace {
    * @param contractAddresses - An array of NFT contract addresses to check ownership for.
    * @beta
    */
-  async checkNftOwnership(
+  checkNftOwnership(
     owner: string,
     contractAddresses: string[]
   ): Promise<boolean> {
-    if (contractAddresses.length === 0) {
-      throw new Error('Must provide at least one contract address');
-    }
-    const response = await this.getNftsForOwner(owner, {
-      contractAddresses,
-      omitMetadata: true
-    });
-    return response.ownedNfts.length > 0;
+    return checkNftOwnership(this.config, owner, contractAddresses);
   }
 
   /**
@@ -379,14 +257,7 @@ export class NftNamespace {
    * @beta
    */
   isSpamContract(contractAddress: string): Promise<boolean> {
-    return requestHttpWithBackoff<IsSpamContractParams, boolean>(
-      this.config,
-      AlchemyApiType.NFT,
-      'isSpamContract',
-      {
-        contractAddress
-      }
-    );
+    return isSpamContract(this.config, contractAddress);
   }
 
   /**
@@ -397,12 +268,7 @@ export class NftNamespace {
    * @beta
    */
   getSpamContracts(): Promise<string[]> {
-    return requestHttpWithBackoff<undefined, string[]>(
-      this.config,
-      AlchemyApiType.NFT,
-      'getSpamContracts',
-      undefined
-    );
+    return getSpamContracts(this.config);
   }
 
   /**
@@ -412,14 +278,7 @@ export class NftNamespace {
    * @beta
    */
   getFloorPrice(contractAddress: string): Promise<GetFloorPriceResponse> {
-    return requestHttpWithBackoff<GetFloorPriceParams, GetFloorPriceResponse>(
-      this.config,
-      AlchemyApiType.NFT,
-      'getFloorPrice',
-      {
-        contractAddress
-      }
-    );
+    return getFloorPrice(this.config, contractAddress);
   }
 
   /**
@@ -437,14 +296,11 @@ export class NftNamespace {
    * @param contractAddress - The contract address of the NFT.
    * @param tokenId - The token id of the NFT.
    */
-  async refreshNftMetadata(
+  refreshNftMetadata(
     contractAddress: string,
     tokenId: BigNumberish
   ): Promise<boolean> {
-    const tokenIdString = BigNumber.from(tokenId!).toString();
-    const first = await this.getNftMetadata(contractAddress, tokenIdString);
-    const second = await refresh(this.config, contractAddress, tokenIdString);
-    return first.timeLastUpdated !== second.timeLastUpdated;
+    return refreshNftMetadata(this.config, contractAddress, tokenId);
   }
 
   /**
@@ -458,189 +314,7 @@ export class NftNamespace {
    * @param contractAddress - The contract address of the NFT collection.
    * @beta
    */
-  async refreshContract(
-    contractAddress: string
-  ): Promise<RefreshContractResult> {
-    const response = await requestHttpWithBackoff<
-      ReingestContractParams,
-      RawReingestContractResponse
-    >(this.config, AlchemyApiType.NFT, 'reingestContract', {
-      contractAddress
-    });
-
-    return {
-      contractAddress: response.contractAddress,
-      refreshState: parseReingestionState(response.reingestionState),
-      progress: response.progress
-    };
+  refreshContract(contractAddress: string): Promise<RefreshContractResult> {
+    return refreshContract(this.config, contractAddress);
   }
-}
-
-async function refresh(
-  config: AlchemyConfig,
-  contractAddress: string,
-  tokenId: BigNumberish
-): Promise<Nft> {
-  const response = await requestHttpWithBackoff<GetNftMetadataParams, RawNft>(
-    config,
-    AlchemyApiType.NFT,
-    'getNFTMetadata',
-    {
-      contractAddress,
-      tokenId: BigNumber.from(tokenId!).toString(),
-      refreshCache: true
-    }
-  );
-  return getNftFromRaw(response, contractAddress);
-}
-
-/**
- * Helper method to convert a NFT response received from Alchemy backend to an
- * SDK NFT type.
- *
- * @internal
- */
-function nftFromGetNftResponse(
-  ownedNft: RawOwnedBaseNft | RawOwnedNft
-): Nft | BaseNft {
-  if (isNftWithMetadata(ownedNft)) {
-    return getNftFromRaw(ownedNft, ownedNft.contract.address);
-  } else {
-    return getBaseNftFromRaw(ownedNft, ownedNft.contract.address);
-  }
-}
-
-/**
- * Helper method to convert a NFT response received from Alchemy backend to an
- * SDK NFT type.
- *
- * @internal
- */
-function nftFromGetNftNftContractResponse(
-  ownedNft: RawContractBaseNft | RawContractNft,
-  contractAddress: string
-): Nft | BaseNft {
-  if (isNftWithMetadata(ownedNft)) {
-    return getNftFromRaw(ownedNft, contractAddress);
-  } else {
-    return getBaseNftFromRaw(ownedNft, contractAddress);
-  }
-}
-
-/** @internal */
-// TODO: more comprehensive type check
-function isNftWithMetadata(response: RawBaseNft | RawNft): response is RawNft {
-  return (response as RawNft).title !== undefined;
-}
-
-/**
- * Flips the `omitMetadata` SDK parameter type to the `withMetadata` parameter
- * required by the Alchemy API. If `omitMetadata` is undefined, the SDK defaults
- * to including metadata.
- *
- * @internal
- */
-function omitMetadataToWithMetadata(
-  omitMetadata: boolean | undefined
-): boolean {
-  return omitMetadata === undefined ? true : !omitMetadata;
-}
-
-function parseReingestionState(reingestionState: string): RefreshState {
-  switch (reingestionState) {
-    case 'does_not_exist':
-      return RefreshState.DOES_NOT_EXIST;
-    case 'already_queued':
-      return RefreshState.ALREADY_QUEUED;
-    case 'in_progress':
-      return RefreshState.IN_PROGRESS;
-    case 'finished':
-      return RefreshState.FINISHED;
-    case 'queued':
-      return RefreshState.QUEUED;
-    case 'queue_failed':
-      return RefreshState.QUEUE_FAILED;
-    default:
-      throw new Error('Unknown reingestion state: ' + reingestionState);
-  }
-}
-
-/**
- * Interface for the `getNftsForNftContract` endpoint. The main difference is
- * that the endpoint has a `startToken` parameter, but the SDK standardizes all
- * pagination parameters to `pageKey`.
- *
- * @internal
- */
-interface GetNftsForNftContractAlchemyParams {
-  contractAddress: string;
-  startToken?: string;
-  withMetadata: boolean;
-}
-
-/**
- * Interface for the `getNfts` endpoint. The main difference is that the
- * endpoint has a `withMetadata` parameter, but the SDK exposes the parameter as
- * `omitMetadata`.
- *
- * @internal
- */
-interface GetNftsAlchemyParams {
-  owner: string;
-  pageKey?: string;
-  contractAddresses?: string[];
-  filters?: string[];
-  withMetadata: boolean;
-}
-
-/**
- * Interface for the `getNftMetadata` endpoint.
- *
- * @internal
- */
-interface GetNftMetadataParams {
-  contractAddress: string;
-  tokenId: string;
-  tokenType?: NftTokenType;
-  refreshCache?: boolean;
-}
-
-/**
- * Interface for the `isSpamContract` endpoint.
- *
- * @internal
- */
-interface IsSpamContractParams {
-  contractAddress: string;
-}
-
-/**
- * Interface for the `getNftContractMetadata` endpoint.
- *
- * @internal
- */
-interface GetContractMetadataParams {
-  contractAddress: string;
-}
-
-/**
- * Interface for the `getOwnersForNftContract` endpoint.
- *
- * @internal
- */
-interface GetOwnersForNftContractAlchemyParams {
-  contractAddress: string;
-}
-
-/**
- * Interface for the `getFloorPrice` endpoint.
- *
- * @internal
- */
-interface GetFloorPriceParams {
-  contractAddress: string;
-}
-
-interface ReingestContractParams {
-  contractAddress: string;
 }
